Proxy remote-auth-gateway on KaiOS
The remote-auth-gateway is more guarded than the default gateway by
Cloudflare, so I'm just hosting my own proxy to bypass this.

The proxy is heavily rate-limited and is provided as a best-effort type
service, so there's no guarantee that it's up 24/7.

Also, secrets exchanged to Discord are encrypted using a locally
generated public/private key pair, so the proxy would see none of that.

The proxy is located at discord-remote-auth.libdb.so.